pyproject.toml
[project]
name = "bitbucket-server"
version = "0.1.0"
description = "BitBucket Server REST API CLI Tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.33.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
]
